Replaced HostOp ALLOC_SEMAPHORE with class SemaphoreGuard.
The class has to be provided by the userland that uses Seoul. The
constructor should grab the lock and the destructor should release it.
Additionally, it has a subclass Semaphore whose default-constructor
creates a semaphore that is passed to the constructor of the guard.

The reason for the change is that the previously used guard-class from
NUL doesn't fit all because it is NOVA specific. And since it doesn't
require a lot of code and it's not that easy to extract a general part,
I think the best way is to let the userland provide the class.